fix: change browser build from IIFE to ESM format
The browser build was previously an IIFE that assigned to a global
XwordParser variable, which caused issues with Vite and other modern
bundlers that expect ES modules. Changed the tsup configuration to
build the browser bundle as an ES module with proper exports.

This fixes import errors when using the library in Vite/TypeScript
projects where the bundler prioritizes the browser field from
package.json exports.